drivers: imu: adis: Rework burst read API
Rework burst read API such that it can be used for devices
which have 32-bit temperature data and 32-bit data counter.
Furthermore, some devices support burst reading with 16-bit LSB
first, while other support burst reading with 16-bit MSB first
in burst data, thus memcpy cannot be used anymore to simply
copy the whole burst data, instead, a burst data structure has
been created which should be populated accordingly by each driver.
The iio_adis generic implementation now populates the iio buffer
using this structure.
Rework the unit tests accordingly.

+++ drivers/imu/iio_adis_internals.h
@@ -185,7 +185,7 @@ struct adis_iio_dev {
 	/** Number of lost samples for the current buffer reading. */
 	uint16_t samples_lost;
 	/** Current data counter for the current buffer reading.*/
-	uint16_t data_cntr;
+	uint32_t data_cntr;
 	/** ADIS sampling frequency. */
 	uint32_t sampling_frequency;
 	/** Current setting for adis burst size. */
